server.py
Server Algorithm For ARP spoofing prevention technique

    The server algorithm, described in Algorithm 2, listens to incoming register messages from the clients, checks the hash code to be 
sure that the message is coming from a trusted host.Users are given only three trials to send the correct hash code. If it fails to send
the correct hash code within the three trials, the server will block this user. The blocking action depends on the addressing chemebeing 
used, for DHCP networks, the MAC address of the user will be added to DHCP deny list. Hence, it will not be able to obtain IP 
configuration from the DHCP server again, for static networks, the server will prevent traffic from this user to reach the server by 
obstructing its IP address.