Client Sniff Program
The algorithm checks the source IP address of the received message to be sure that it is coming from the trusted server. It only 
accepts the IP and MAC addresses encapsulated in the message if the key is correct.In order to work in DHCP networks where IP and
MAC mappings are frequently changing, the algorithm searches for the MAC encapsulated in the message. If matched map is found, it 
will be changed to the new mapping. Otherwise a new mapping will be added. Finally if any of the conditions are not met, the algorithm will discard the message and return to listen for another message from the server.