Client Program
The client algorithm described in Algorithm 1 adds static entry for the server in the client ARP cache to avoid the rogue server threat.
Furthermore, it obtains the user IP and MAC address automatically to make the user has no opportunity to send fake information to the 
server. This program packs registration message and send it to Server.