Rename server.py to server_program.py
It is a program that can be deployed on any device in the network. It can also be installed on a dedicated server, and has the following
Functions: 
        Receive register messages from the ARP clients.
        Verify that the message is coming from a trusted user.
        Make use of the IP and MAC pairs encapsulated within the register message to create a list of trusted users in the network.
        Send broadcast update message to notify them that a new user has come to the network.
        Send register response message to the new users.
        Take the proper action regarding users who try to violate the protocol security rules.

ARP_Server Algorithm:
         The server algorithm, described in Algorithm 2, listens to incoming register messages from the clients, checks the hash code to 
be sure that the message is coming from a trusted host. Users are given only three trials to send the correct hash code. If it fails to 
send the correct hash code within the three trials, the server will block this user. The blocking action depends on the addressing 
schema being used, for DHCP networks, the MAC address of the user will be added to DHCP deny list. Hence, it will not be able to 
obtain IP configuration from the DHCP server again, for static networks, the server will prevent traffic from this user to reach the server 
by obstructing its IP address.